+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes
 -----
 
 - Differentiate webmap intial and contstraining extents.
+- New OGC API Feature service with read and write support.
 
 
 4.4.0